requirements.txt
sanic # Ta race
python-dotenv
pdfkit
wkhtmltopdf
jinja2